refactor : Domain model 분리
