+++ domain/src/main/java/com/example/domain/model/Coordinate.kt
@@ -1,0 +1,6 @@
+package com.example.domain.model
+
+data class Coordinate(
+    val longitude: Double,
+    val latitude: Double,
+)
