+++ domain/src/main/java/com/example/domain/model/OpeningHours.kt
@@ -1,0 +1,6 @@
+package com.example.domain.model
+
+data class OpeningHours(
+    val open: TimeInfo,
+    val close: TimeInfo,
+)
