+++ domain/src/main/java/com/example/domain/model/TimeInfo.kt
@@ -1,0 +1,7 @@
+package com.example.domain.model
+
+data class TimeInfo(
+    val day: String,
+    val hour: Int,
+    val minute: Int,
+)
